backend/Rejestracja/requirements.txt
PyGreSQL
python-dotenv
fastapi[standard]
uvicorn
pydantic
passlib[bcrypt]
pyjwt
apscheduler